Consolidate setupinfo into setup
